Rename from planning to services overview
