+++ app/views/base/organizations/_navbar.html.slim
@@ -14,7 +14,7 @@ nav.navbar.navbar-vertical.navbar-expand-xl.navbar-light
         span.fas.fa-chart-pie
       = render 'base/navbar_list_item', link: { t('.civil_servants') => organizations_civil_servants_path } do
         span.fas.fa-tools
-      = render 'base/navbar_list_item', link: { t('.planning') => organizations_planning_path } do
+      = render 'base/navbar_list_item', link: { t('.services_overview') => organizations_services_path } do
         span.fas.fa-calendar-alt
       = render 'base/navbar_list_item', link: { t('.service_specifications') => organizations_service_specifications_path } do
         span.fas.fa-file
