--- spec/views/organizations/planning/index.html.slim_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe 'planning/index.html.slim', type: :view do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
