+++ spec/views/organizations/services/index.html.slim_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe 'organizations/services/index.html.slim', type: :view do
+  before { render }
+
+  it 'has preact container' do
+    expect(rendered).to include 'id="planning-app"'
+  end
+end
